Added rc-local file, startup scripts

+++ host/rsync/rsync_beaglebone.sh
@@ -12,21 +12,31 @@ RSYNC_LOCAL="/root"
 RSYNC_PORT="873"
 PROJECT=$1
 
+
 # Proceed if a project was requested
 if [ ! -z ${PROJECT} ]; then
     # Check whether Rsync Server is available - First item in rsync.conf must be "online"
     SYNC_AVAILABLE=`rsync -n $RSYNC_SERVER::`;
     if [ "${SYNC_AVAILABLE%% *}" = "online" ]; then
         # Server is available. Check if there are updates
-        UPDATES=`rsync -ainO $RSYNC_SERVER::$PROJECT $RSYNC_LOCAL/$PROJECT`;
+        if [ "${PROJECT}" = "etc-folder" ]; then
+            UPDATES=`rsync -ainO $RSYNC_SERVER::$PROJECT /etc`;
+        else
+            UPDATES=`rsync -ainO $RSYNC_SERVER::$PROJECT $RSYNC_LOCAL/$PROJECT`;
+        fi
 
         # No updates available
         if [ -z "$UPDATES" ]; then
             echo "No updates found.";
+            exit 1
 
         # There are updates for the project, sync files and build libraries
         else
-            rsync -a --delete-after $RSYNC_SERVER::$PROJECT $RSYNC_LOCAL/$PROJECT > /tmp/rsync.log;
+            if [ "${PROJECT}" = "etc-folder" ]; then
+                rsync -a $RSYNC_SERVER::$PROJECT /etc > /tmp/rsync.log;
+            else
+                rsync -a --delete-after $RSYNC_SERVER::$PROJECT $RSYNC_LOCAL/$PROJECT > /tmp/rsync.log;
+            fi
             if [ $? -eq 0 ]; then
                 # If project is listed below, build libraries as well
                 if [ "$PROJECT" = "pru-serial485" ] || [ "$PROJECT" = "counting-pru" ]; then
